refactor: remove all paver code related to python unit tests
