+++ docs/guides/testing/testing.rst
@@ -220,8 +220,7 @@ The report is then saved in reports/xmodule/cover/index.html
 To run tests for stub servers, for example for `YouTube stub server`_, you can
 run one of these commands::
 
-    paver test_system -s cms -t common/djangoapps/terrain/stubs/tests/test_youtube_stub.py
-    pytest common/djangoapps/terrain/stubs/tests/test_youtube_stub.py
+    pytest --ds=cms.env.test common/djangoapps/terrain/stubs/tests/test_youtube_stub.py
 
 .. _YouTube stub server: https://github.com/openedx/edx-platform/blob/master/common/djangoapps/terrain/stubs/tests/test_youtube_stub.py
