+++ pavelib/js_test.py
@@ -3,6 +3,8 @@ Javascript test tasks
 """
 
 
+import os
+import re
 import sys
 
 from paver.easy import cmdopts, needs, task
@@ -10,6 +12,12 @@ from paver.easy import cmdopts, needs, task
 from pavelib.utils.envs import Env
 from pavelib.utils.test.suites import JestSnapshotTestSuite, JsTestSuite
 from pavelib.utils.timer import timed
+from paver.easy import cmdopts, needs, sh, task
+
+try:
+    from pygments.console import colorize
+except ImportError:
+    colorize = lambda color, text: text
 
 __test__ = False  # do not collect
 
@@ -91,3 +99,47 @@ def test_js_dev(options):
     """
     options.mode = 'dev'
     test_js(options)
+
+
+@task
+@needs('pavelib.prereqs.install_coverage_prereqs')
+@cmdopts([
+    ("compare-branch=", "b", "Branch to compare against, defaults to origin/master"),
+], share_with=['coverage'])
+@timed
+def diff_coverage(options):
+    """
+    Build the diff coverage reports
+    """
+    compare_branch = options.get('compare_branch', 'origin/master')
+
+    # Find all coverage XML files (both Python and JavaScript)
+    xml_reports = []
+
+    for filepath in Env.REPORT_DIR.walk():
+        if bool(re.match(r'^coverage.*\.xml$', filepath.basename())):
+            xml_reports.append(filepath)
+
+    if not xml_reports:
+        err_msg = colorize(
+            'red',
+            "No coverage info found.  Run `paver test` before running "
+            "`paver coverage`.\n"
+        )
+        sys.stderr.write(err_msg)
+    else:
+        xml_report_str = ' '.join(xml_reports)
+        diff_html_path = os.path.join(Env.REPORT_DIR, 'diff_coverage_combined.html')
+
+        # Generate the diff coverage reports (HTML and console)
+        # The --diff-range-notation parameter is a workaround for https://github.com/Bachmann1234/diff_cover/issues/153
+        sh(
+            "diff-cover {xml_report_str} --diff-range-notation '..' --compare-branch={compare_branch} "
+            "--html-report {diff_html_path}".format(
+                xml_report_str=xml_report_str,
+                compare_branch=compare_branch,
+                diff_html_path=diff_html_path,
+            )
+        )
+
+        print("\n")
