+++ scripts/all-tests.sh
@@ -16,18 +16,6 @@ source scripts/thresholds.sh
 XSSLINT_THRESHOLDS=`cat scripts/xsslint_thresholds.json`
 export XSSLINT_THRESHOLDS=${XSSLINT_THRESHOLDS//[[:space:]]/}
 
-doCheckVars() {
-    if [ -n "$CIRCLECI" ] ; then
-        SCRIPT_TO_RUN=scripts/circle-ci-tests.sh
-
-    elif [ -n "$JENKINS_HOME" ] ; then
-        source scripts/jenkins-common.sh
-        SCRIPT_TO_RUN=scripts/generic-ci-tests.sh
-    fi
-}
-
-# Determine the CI system for the environment
-doCheckVars
 
 # Run appropriate CI system script
 if [ -n "$SCRIPT_TO_RUN" ] ; then
