+++ scripts/paver_autocomplete.sh
@@ -56,16 +56,6 @@ _paver()
 
     if [[ -n $subcmd ]]
     then
-        case $subcmd in
-            test_system)
-
-                _test_system_args
-                if [[ -n $COMPREPLY ]]
-                then
-                    return 0
-                fi
-                ;;
-        esac
 
         if [[ ${#COMP_WORDS[*]} == 3 ]]
         then
@@ -86,22 +76,6 @@ _paver()
     then
         COMPREPLY=( $(compgen -W "${cmds}" -- "$cur") )
     fi
-}
-
-_test_system_args()
-{
-        local cur prev
-        cur="${COMP_WORDS[COMP_CWORD]}"
-        prev="${COMP_WORDS[COMP_CWORD - 1]}"
-
-        case "$prev" in
-            -s|--system)
-                COMPREPLY=( $(compgen -W "lms cms" -- "$cur") )
-                return 0
-                ;;
-            *)
-                ;;
-        esac
 }
 
 # Assign the auto-completion function for our command.
